Add initial dataset as a result of ChatGPT running the prompt
The dataset is invalid and needs to be cleaned up. It should be an array of arrays of objects. Carefule with multiple possible keys (ex: has both instructiune and sarcina) - but this will be checked and replaced in a script after the dataset is valid so it could be imported